Create StatusItemScroller
